[SR-py] simplify ms/fr/generate
