+++ portfolio_10k_net_income_report.txt
@@ -1,0 +1,6 @@
+1. NVDA: The COVID-19 pandemic has affected and could continue to have a material adverse impact on the financial condition and results of operations. It's impacted our workforce and operations and those of our customers, partners, vendors and suppliers. COVID-19-related disruptions have created and may continue to create supply chain and logistics constraints. 
+2. AAPL: The COVID-19 pandemic has significantly curtailed global economic activity and caused significant volatility and disruption in global financial markets. It has adversely affected and could in the future materially adversely impact the business, results of operations, financial.
+3. JNJ: The future operations may depend on many factors which cannot be predicted with confidence, including the duration of an outbreak and impact of variants. A surge in COVID-19 could result in new mandates and prolonged restrictive measures that could adversely impact operations. 
+4. PEP: The novel coronavirus (COVID-19) pandemic has contributed to adverse conditions such as the inflationary cost environment, adverse weather conditions, supply chain disruptions and labor shortages, continued into fiscal 2023.
+5. UNH: The impact of COVID-19 could increase our medical and administrative costs, and result in reduced capacity to handle demand for care or impact our business operations. COVID-19 has materially impacted results of operations in previous periods. 
+6. AMZN: Macroeconomic factors like the prolonged COVID-19 pandemic, global supply chain constraints, and global economic and geopolitical developments, have direct and indirect impacts on the results of operations. These factors contributed to increases in our operating costs during 2022.
