+++ python/src/semantic_retrieval/aiconfigs/py-completion-gen-aiconfig_aiconfig.json
@@ -10,7 +10,11 @@
         "frequency_penalty": 0
       }
     },
-    "parameters": {}
+    "parameters": {
+      "retrieval_query": "overall cash flow",
+      "structure_prompt": "Numbered list, one security per list item,",
+      "data_extraction_prompt": "covid 19 impact"
+    }
   },
   "prompts": [
     {
