+++ python/src/semantic_retrieval/examples/financial_report/access_control/iam_simulation_db.json
@@ -5,8 +5,7 @@
   },
   "access_10ks": {
     "advisor/jonathan": [
-      "AAPL",
-      "PEP"
+      "AAPL", "PEP", "AMZN", "NVDA", "UNH", "JNJ"
     ]
   }
 }
