+++ python/src/semantic_retrieval/examples/financial_report/lib/common.py
@@ -1,0 +1,10 @@
+from typing import Dict, NewType
+import pandas as pd
+
+PortfolioData = NewType("PortfolioData", Dict[str, int])
+
+
+def portfolio_df_to_dict(df: pd.DataFrame) -> PortfolioData:
+    return PortfolioData(
+        df.set_index("Company").astype(float).fillna(0).query("Shares > 0")["Shares"].to_dict()
+    )
