+++ report.out
@@ -1,0 +1,11 @@
+1. NVDA: The COVID-19 pandemic has and could continue to affect our financial condition and results of operations. Such impacts include workforce and operational disturbances, supply chain and logistics constraints, and global restrictions. 
+
+2. AAPL:  The COVID-19 pandemic has curtailed global economic activity, caused volatility in global financial markets, and resulted in business operation restrictions. These scenarios have adversely affected the business, results of operations and financial.
+
+3. JNJ: The impact of the COVID-19 pandemic on the Company's future operations is dependent on unpredictable factors such as the duration of an outbreak and impact of variants. A surge in COVID-19 could result in the imposition of new mandates and impact the Company's operations.
+
+4. PEP: The COVID-19 pandemic along with other factors such as inflation, adverse weather conditions, supply chain disruptions and labor shortages, have continued to affect the company in fiscal 2023.
+
+5. UNH: The impact of COVID-19 includes potential increases in our medical and administrative costs, reduced demand for our services, and reduced capacity to handle demand for care. The pandemic has significantly affected operations in previous periods.
+
+6. AMZN: Macroeconomic factors including the prolonged COVID-19 pandemic, global supply chain constraints, global economic and geopolitical developments, have direct and indirect impacts on the results of operations. These factors contributed to increases in operating costs during 2022.
